Server waiting procedure talks when --debug is set.
Silent errors are evil because users have to guess what is going
wrong or what they did wrong. With this patch we give some hint to the
user about what is going one while the master process is trying to ping
the server being started. A timeout is also added so that all processes
(including the daemon process) are shutdown properly after a given
amount of time.